refactored the notification hangar into a library
